app/requirements.txt
torchgeo[all]
torchdiffeq
neurodiffeq